+++ R/clean_parameters.R
@@ -361,8 +361,13 @@ clean_parameters.stanmvreg <- function(x, ...) {
 
   cor_sd <- grepl("(sd_|cor_)(.*)", out$Cleaned_Parameter)
   if (any(cor_sd)) {
-    out$Cleaned_Parameter[cor_sd] <- gsub("(sd_|cor_)(.*)__(.*)", "\\2_\\3", out$Cleaned_Parameter[cor_sd])
-    out$Group[cor_sd] <- "SD/Cor"
+    out$Cleaned_Parameter[cor_sd] <- gsub("^(sd_|cor_)(.*?)__(.*)", "\\3", out$Parameter[cor_sd], perl = TRUE)
+    out$Group[cor_sd] <- paste("SD/Cor:", gsub("^(sd_|cor_)(.*?)__(.*)", "\\2", out$Parameter[cor_sd], perl = TRUE))
+    # replace "__" by "~"
+    cor_only <- grepl("^cor_", out$Parameter[cor_sd])
+    if (any(cor_only)) {
+      out$Cleaned_Parameter[which(cor_sd)[cor_only]] <- sub("__", " ~ ", out$Cleaned_Parameter[which(cor_sd)[cor_only]])
+    }
   }
 
   # extract group-names from random effects and clean random effects
@@ -437,6 +442,21 @@ clean_parameters.stanmvreg <- function(x, ...) {
       out$Cleaned_Parameter[rand_intercepts]
     )
     out$Group[rand_intercepts] <- sprintf("Intercept: %s", re_grp_level)
+  }
+
+
+  # correlation and sd
+
+  cor_sd <- grepl("^Sigma\\[(.*)", out$Cleaned_Parameter)
+  if (any(cor_sd)) {
+    parm1 <- gsub("^Sigma\\[(.*):(.*),(.*)\\]", "\\2", out$Parameter[cor_sd], perl = TRUE)
+    parm2 <- gsub("^Sigma\\[(.*):(.*),(.*)\\]", "\\3", out$Parameter[cor_sd], perl = TRUE)
+    out$Cleaned_Parameter[which(cor_sd)] <- parm1
+    rand_cor <- parm1 != parm2
+    if (any(rand_cor)) {
+      out$Cleaned_Parameter[which(cor_sd)[rand_cor]] <- paste0(parm1[rand_cor], " ~ ", parm2[rand_cor])
+    }
+    out$Group[cor_sd] <- paste("SD/Cor:", gsub("^Sigma\\[(.*):(.*),(.*)\\]", "\\1", out$Parameter[cor_sd], perl = TRUE))
   }