+++ R/find_parameters.R
@@ -677,6 +677,7 @@ find_parameters.BFBayesFactor <- function(x, effects = c("all", "fixed", "random
 
     params <- colnames(posteriors)
     vars <- find_variables(x, effects = "all")
+    interactions <- find_interactions(x)
     dat <- get_data(x)
 
     if ("conditional" %in% names(vars)) {
@@ -687,6 +688,13 @@ find_parameters.BFBayesFactor <- function(x, effects = c("all", "fixed", "random
           i
         }
       }))
+    }
+
+    # add interaction terms to conditional
+    if ("conditional" %in% names(interactions)) {
+      for (i in interactions$conditional) {
+        conditional <- c(conditional, params[grepl(paste0("^\\Q", i, "\\E"), params)])
+      }
     }
 
     if ("random" %in% names(vars)) {
@@ -739,18 +747,22 @@ find_parameters.brmsfit <- function(x, effects = c("all", "fixed", "random"), co
 
   cond <- fe[grepl(pattern = "^(b_|bs_|bsp_|bcs_)(?!zi_)(.*)", fe, perl = TRUE)]
   zi <- fe[grepl(pattern = "^(b_zi_|bs_zi_|bsp_zi_|bcs_zi_)", fe, perl = TRUE)]
-  rand <- fe[grepl(pattern = "(?!.*__zi)(?=.*r_)", fe, perl = TRUE) & !grepl(pattern = "^prior_", fe, perl = TRUE)]
+  rand <- fe[grepl(pattern = "(?!.*__zi)(?=.*^r_)", fe, perl = TRUE) & !grepl(pattern = "^prior_", fe, perl = TRUE)]
   randzi <- fe[grepl(pattern = "^r_(.*__zi)", fe, perl = TRUE)]
   simo <- fe[grepl(pattern = "^simo_", fe, perl = TRUE)]
   smooth_terms <- fe[grepl(pattern = "^sds_", fe, perl = TRUE)]
   priors <- fe[grepl(pattern = "^prior_", fe, perl = TRUE)]
   sigma <- fe[grepl(pattern = "^sigma_", fe, perl = TRUE)]
+  rand_sd <- fe[grepl(pattern = "(?!.*_zi)(?=.*^sd_)", fe, perl = TRUE)]
+  randzi_sd <- fe[grepl(pattern = "^sd_(.*_zi)", fe, perl = TRUE)]
+  rand_cor <- fe[grepl(pattern = "(?!.*_zi)(?=.*^cor_)", fe, perl = TRUE)]
+  randzi_cor <- fe[grepl(pattern = "^cor_(.*_zi)", fe, perl = TRUE)]
 
   l <- .compact_list(list(
     conditional = cond,
-    random = rand,
+    random = c(rand, rand_sd, rand_cor),
     zero_inflated = zi,
-    zero_inflated_random = randzi,
+    zero_inflated_random = c(randzi, randzi_sd, randzi_cor),
     simplex = simo,
     smooth_terms = smooth_terms,
     sigma = sigma,
@@ -879,11 +891,12 @@ find_parameters.stanreg <- function(x, effects = c("all", "fixed", "random"), co
 
   cond <- fe[grepl(pattern = "^(?!(b\\[|sigma|Sigma))", fe, perl = TRUE) & .grep_non_smoothers(fe)]
   rand <- fe[grepl(pattern = "^b\\[", fe, perl = TRUE)]
+  rand_sd <- fe[grepl(pattern = "^Sigma\\[", fe, perl = TRUE)]
   smooth_terms <- fe[grepl(pattern = "^smooth_sd", fe, perl = TRUE)]
 
   l <- .compact_list(list(
     conditional = cond,
-    random = rand,
+    random = c(rand, rand_sd),
     smooth_terms = smooth_terms
   ))
 
@@ -1142,6 +1155,34 @@ find_parameters.wbgee <- find_parameters.wbm
 
 
 # Other models -----------------------------------
+
+
+#' @importFrom stats na.omit coef
+#' @export
+find_parameters.manova <- function(x, flatten = FALSE, ...) {
+  out <- list(conditional = .remove_backticks_from_string(rownames(stats::na.omit(stats::coef(x)))))
+
+  if (flatten) {
+    unique(unlist(out))
+  } else {
+    out
+  }
+}
+
+
+#' @rdname find_parameters
+#' @export
+find_parameters.averaging <- function(x, component = c("conditional", "full"), flatten = FALSE, ...) {
+  component <- match.arg(component)
+  cf <- stats::coef(x, full = component == "full")
+  out <- list(conditional = .remove_backticks_from_string(names(cf)))
+
+  if (flatten) {
+    unique(unlist(out))
+  } else {
+    out
+  }
+}
 
 
 #' @export